requirements.txt
requests
beautifulsoup4
PyGithub
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pandas
selenium